Refactored the code for DNSSD portion to discoverer class
