+++ lib/whipped-cream/cli.rb
@@ -1,5 +1,4 @@
 require 'thor'
-require 'dnssd'
 
 module WhippedCream
   # The CLI gets invoked from the binary, and encapsulates all user interaction
@@ -50,76 +49,13 @@ module WhippedCream
 
     desc "discover", "Discovers any whipped-cream servers on the local network"
     def discover
-      services = browse_services('_whipped-cream._tcp.')
-
-      services.each do |service|
-        host = resolve_service(service)
-
-        puts "#{host[:address]}:#{host[:port]}\t#{host[:name]}"
-      end
-
-      services
+      Discoverer.discover '_whipped-cream._tcp.'
     end
 
     no_tasks do
       def resolve_plugin(name)
         name += '.rb' unless name.split('.').last == 'rb'
         name
-      end
-
-      def browse_services(service_type)
-        services = []
-
-        begin
-          timeout 5 do
-            DNSSD::Service.new.browse(service_type) do |reply|
-              services << reply if reply.flags.add?
-              break unless reply.flags.more_coming?
-            end
-          end
-        rescue Timeout::Error
-          log "ERROR: #browse_services took more than 5 seconds."
-        end
-
-        services
-      end
-
-      def resolve_service(service)
-        host = {}
-
-        begin
-          timeout 5 do
-            DNSSD::Service.new.resolve(service) do |reply|
-              address = get_ipv4_address(reply)
-
-              host[:name]    = "#{reply.name}"
-              host[:address] = "#{address}"
-              host[:port]    = "#{reply.port}"
-              break unless reply.flags.more_coming?
-            end
-          end
-        rescue Timeout::Error
-          log "ERROR: #resolve_service took more than 5 seconds."
-        end
-
-        host
-      end
-
-      def get_ipv4_address(reply)
-        address = nil
-
-        begin
-          timeout 5 do
-            DNSSD::Service.new.getaddrinfo(reply.target, 1) do |addrinfo|
-              address = addrinfo.address
-              break unless addrinfo.flags.more_coming?
-            end
-          end
-        rescue Timeout::Error
-          log "ERROR: #get_ipv4_address took more than 5 seconds."
-        end
-
-        address
       end
     end
   end
